refactor: 기능에 맞게 api 재배치 및 불필요한 코드 제거
