--- src/main/java/com/gotcha/server/global/config/QueryDslConfig.java
@@ -1,18 +1,0 @@
-package com.gotcha.server.global.config;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.PersistenceContext;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class QueryDslConfig {
-    @PersistenceContext
-    private EntityManager em;
-
-    @Bean
-    public JPAQueryFactory jpaQueryFactory() {
-        return new JPAQueryFactory(em);
-    }
-}
