+++ src/main/java/com/gotcha/server/mail/config/EmailConfig.java
@@ -1,9 +1,8 @@
-package com.gotcha.server.global.config;
+package com.gotcha.server.mail.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.PropertySource;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.mail.javamail.JavaMailSenderImpl;
