+++ src/main/java/com/gotcha/server/member/dto/response/TodayInterviewResponse.java
@@ -1,4 +1,4 @@
-package com.gotcha.server.applicant.dto.response;
+package com.gotcha.server.member.dto.response;
 
 public record TodayInterviewResponse(long count) {
